deep-learning-model/requirements.txt
tensorflow>=2.12
flask>=2.3
requests>=2.31
numpy>=1.24
pillow>=10.0
